CORE-4874 Removed Security Manager permissions for "net.corda.testing*" used by cordapps as "net.corda*" is reserved for platform implementation classes. Refactored bundles testing-cats and testing-dogs to use package "net.cordapp.testing" instead of "net.corda.testing"
