+++ components/db/db-connection-manager-impl/src/integrationTest/kotlin/net/corda/db/connection/manager/impl/tests/DbAdminTest.kt
@@ -20,7 +20,7 @@ import net.corda.lifecycle.LifecycleCoordinatorFactory
 import net.corda.orm.EntityManagerConfiguration
 import net.corda.orm.impl.EntityManagerFactoryFactoryImpl
 import net.corda.orm.impl.JpaEntitiesRegistryImpl
-import net.corda.testing.bundles.cats.Cat
+import net.cordapp.testing.bundles.cats.Cat
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.Assumptions
