+++ components/virtual-node/entity-processor-service-impl/src/integrationTest/kotlin/net/corda/entityprocessor/impl/tests/CpbEntityTests.kt
@@ -58,7 +58,7 @@ class CpbEntityTests {
 
         // does contain packages from cpks
         assertThat(entities).containsAll(
-            listOf("net.corda.testing.bundles.dogs.Dog", "net.corda.testing.bundles.cats.Cat", "net.corda.testing.bundles.cats.Owner")
+            listOf("net.cordapp.testing.bundles.dogs.Dog", "net.cordapp.testing.bundles.cats.Cat", "net.cordapp.testing.bundles.cats.Owner")
         )
     }
 }
