+++ components/virtual-node/entity-processor-service-impl/src/integrationTest/kotlin/net/corda/entityprocessor/impl/tests/helpers/SandboxHelper.kt
@@ -15,9 +15,9 @@ import java.util.*
  * twice, with a separate class loader.
  */
 object SandboxHelper {
-    const val DOG_CLASS_NAME = "net.corda.testing.bundles.dogs.Dog"
-    const val CAT_CLASS_NAME = "net.corda.testing.bundles.cats.Cat"
-    const val CAT_KEY_CLASS_NAME = "net.corda.testing.bundles.cats.CatKey"
+    const val DOG_CLASS_NAME = "net.cordapp.testing.bundles.dogs.Dog"
+    const val CAT_CLASS_NAME = "net.cordapp.testing.bundles.cats.Cat"
+    const val CAT_KEY_CLASS_NAME = "net.cordapp.testing.bundles.cats.CatKey"
 
 
     fun SandboxGroup.getDogClass(): Class<*> {
@@ -33,7 +33,7 @@ object SandboxHelper {
     }
 
     fun SandboxGroup.getOwnerClass(): Class<*> {
-        return this.loadClassFromMainBundles("net.corda.testing.bundles.cats.Owner")
+        return this.loadClassFromMainBundles("net.cordapp.testing.bundles.cats.Owner")
     }
 
     data class Box(val instance: Any, val id: UUID)
