+++ testing/bundles/testing-cats/src/main/java/net/cordapp/testing/bundles/cats/package-info.java
@@ -1,4 +1,4 @@
 @Export
-package net.corda.testing.bundles.dogs;
+package net.cordapp.testing.bundles.cats;
 
 import org.osgi.annotation.bundle.Export;
