+++ testing/cpbs/flow-worker-dev/src/main/kotlin/net/cordapp/flowworker/development/testflows/PersistenceFlow.kt
@@ -2,7 +2,7 @@ package net.cordapp.flowworker.development.testflows
 
 import java.time.Instant
 import java.util.UUID
-import net.corda.testing.bundles.dogs.Dog
+import net.cordapp.testing.bundles.dogs.Dog
 import net.corda.v5.application.flows.CordaInject
 import net.corda.v5.application.flows.RPCRequestData
 import net.corda.v5.application.flows.RPCStartableFlow
